# backend/rest/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "GOLSMEDIA API",
    "version": "1.0.0",
    "title": "GOLSMEDIA API",
    "contact": {
      "name": "GOLSMEDIA",
      "url": "https://realworld.io"
    },
    "license": {
      "name": "Alex License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "securityDefinitions": {
    "Token": {
      "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token: xxxxxx.yyyyyyy.zzzzzz\n    \n",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/news/feed": {
      "get": {
        "summary": "Get recent news from users you follow",
        "description": "Get most recent news from users you follow. Use query parameters to limit. Auth is required",
        "tags": [
          "News"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "GetNewsFeed",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of news returned (default is 20)",
            "required": false,
            "default": 20,
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset/skip number of news (default is 0)",
            "required": false,
            "default": 0,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MultipleNewsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
    "/news": {
      "get": {
        "summary": "Get recent news globally",
        "description": "Get most recent news globally. Use query parameters to filter results. Auth is optional",
        "tags": [
          "News"
        ],
        "operationId": "GetNews",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "description": "Filter by tag",
            "required": false,
            "type": "string"
          },
          {
            "name": "author",
            "in": "query",
            "description": "Filter by author (username)",
            "required": false,
            "type": "string"
          },
          {
            "name": "favorited",
            "in": "query",
            "description": "Filter by favorites of a user (username)",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit number of news returned (default is 20)",
            "required": false,
            "default": 20,
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset/skip number of news (default is 0)",
            "required": false,
            "default": 0,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK,",
            "schema": {
              "$ref": "#/definitions/MultipleNewsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "post": {
        "summary": "Create an news",
        "description": "Create an news. Auth is required",
        "tags": [
          "News"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "CreateNew",
        "parameters": [
          {
            "name": "news",
            "in": "body",
            "required": true,
            "description": "News to create",
            "schema": {
              "$ref": "#/definitions/NewNewsRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SingleNewsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
    "/news/{slug}": {
      "get": {
        "summary": "Get an news",
        "description": "Get an news. Auth not required",
        "tags": [
          "News"
        ],
        "operationId": "Getnews",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the new to get",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SingleNewsResponse"
            }
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "put": {
        "summary": "Update an news",
        "description": "Update an news. Auth is required",
        "tags": [
          "News"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "UpdateNews",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the new to update",
            "type": "string"
          },
          {
            "name": "new",
            "in": "body",
            "required": true,
            "description": "New to update",
            "schema": {
              "$ref": "#/definitions/UpdateNewsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SingleNewsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an new",
        "description": "Delete an new. Auth is required",
        "tags": [
          "News"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "DeleteNews",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the new to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
    "/news/{slug}/comments": {
      "get": {
        "summary": "Get comments for an news",
        "description": "Get the comments for an news. Auth is optional",
        "tags": [
          "Comments"
        ],
        "operationId": "GetNewsComments",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the news that you want to get comments for",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MultipleCommentsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "post": {
        "summary": "Create a comment for an news",
        "description": "Create a comment for an news. Auth is required",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "CreateNewsComment",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the news that you want to create a comment for",
            "type": "string"
          },
          {
            "name": "comment",
            "in": "body",
            "required": true,
            "description": "Comment you want to create",
            "schema": {
              "$ref": "#/definitions/NewCommentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SingleCommentResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
    "/news/{slug}/comments/{id}": {
      "delete": {
        "summary": "Delete a comment for an new",
        "description": "Delete a comment for an new. Auth is required",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "DeleteNewsComment",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the new that you want to delete a comment for",
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the comment you want to delete",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
    "/news/{slug}/favorite": {
      "post": {
        "summary": "Favorite an new",
        "description": "Favorite an new. Auth is required",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "CreateNewFavorite",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the new that you want to favorite",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SingleNewsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      },
      "delete": {
        "summary": "Unfavorite an new",
        "description": "Unfavorite an new. Auth is required",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "Token": []
          }
        ],
        "operationId": "DeleteNewsFavorite",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug of the news that you want to unfavorite",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SingleNewsResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/GenericErrorModel"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Profile": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "image"
      ]
    },
    "News": {
      "type": "object",
      "properties": {
        "slug": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "tagList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdNw": {
          "type": "string",
          "format": "date-time"
        },
        "updatedNw": {
          "type": "string",
          "format": "date-time"
        },
        "favorited": {
          "type": "boolean"
        },
        "favoritesCount": {
          "type": "integer"
        },
        "author": {
          "$ref": "#/definitions/Profile"
        }
      },
      "required": [
        "slug",
        "title",
        "description",
        "body",
        "tagList",
        "createdNw",
        "updatedNw",
        "favorited",
        "favoritesCount",
        "author"
      ]
    },
    "SingleNewsResponse": {
      "type": "object",
      "properties": {
        "news": {
          "$ref": "#/definitions/News"
        }
      },
      "required": [
        "news"
      ]
    },
    "MultipleNewsResponse": {
      "type": "object",
      "properties": {
        "news": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/News"
          }
        },
        "newsCount": {
          "type": "integer"
        }
      },
      "required": [
        "news",
        "newsCount"
      ]
    },
    "NewNews": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "tagList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "title",
        "description",
        "body"
      ]
    },
    "NewNewsRequest": {
      "type": "object",
      "properties": {
        "news": {
          "$ref": "#/definitions/NewNews"
        }
      },
      "required": [
        "news"
      ]
    },
    "UpdateNews": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "body": {
          "type": "string"
        }
      }
    },
    "UpdateNewsRequest": {
      "type": "object",
      "properties": {
        "news": {
          "$ref": "#/definitions/UpdateNews"
        }
      },
      "required": [
        "news"
      ]
    },
    "Comment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "body": {
          "type": "string"
        },
        "author": {
          "$ref": "#/definitions/Profile"
        }
      },
      "required": [
        "id",
        "createdAt",
        "updatedAt",
        "body",
        "author"
      ]
    },
    "SingleCommentResponse": {
      "type": "object",
      "properties": {
        "comment": {
          "$ref": "#/definitions/Comment"
        }
      },
      "required": [
        "comment"
      ]
    },
    "MultipleCommentsResponse": {
      "type": "object",
      "properties": {
        "comments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comment"
          }
        }
      },
      "required": [
        "comments"
      ]
    },
    "NewComment": {
      "type": "object",
      "properties": {
        "body": {
          "type": "string"
        }
      },
      "required": [
        "body"
      ]
    },
    "NewCommentRequest": {
      "type": "object",
      "properties": {
        "comment": {
          "$ref": "#/definitions/NewComment"
        }
      },
      "required": [
        "comment"
      ]
    },
...
    "GenericErrorModel": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "object",
          "properties": {
            "body": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "body"
          ]
        }
      },
      "required": [
        "errors"
      ]
    }
  }
}